requirements.txt
fastapi
uvicorn[standard]
PyJWT>=2.8.0
cryptography
httpx
PyGithub
pyyaml
flask
flask-cors
boto3
botocore
python-dotenv
